requirements.txt
google-generativeai==0.3.2
python-dotenv==1.0.1
streamlit==1.32.2
pandas==2.2.1
